+++ src/RoadRegistry.Product.Projections/RoadSegmentRecordProjection.cs
@@ -174,7 +174,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<ProductContext>
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentModified.Id);
         if (roadSegmentRecord == null)
         {
-            throw new InvalidOperationException($"RoadNodeRecord with id {roadSegmentModified.Id} is not found");
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentModified.Id} is not found");
         }
 
         roadSegmentRecord.Id = roadSegmentModified.Id;
@@ -220,7 +220,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<ProductContext>
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentAttributesModified.Id);
         if (roadSegmentRecord == null)
         {
-            throw new InvalidOperationException($"RoadNodeRecord with id {roadSegmentAttributesModified.Id} is not found");
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentAttributesModified.Id} is not found");
         }
 
         roadSegmentRecord.Id = roadSegmentAttributesModified.Id;
@@ -285,7 +285,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<ProductContext>
         var roadSegmentRecord = await context.RoadSegments.FindAsync(segment.Id);
         if (roadSegmentRecord == null)
         {
-            throw new InvalidOperationException($"RoadNodeRecord with id {segment.Id} is not found");
+            throw new InvalidOperationException($"RoadSegmentRecord with id {segment.Id} is not found");
         }
 
         roadSegmentRecord.Id = segment.Id;
@@ -314,6 +314,10 @@ public class RoadSegmentRecordProjection : ConnectedProjection<ProductContext>
     private static async Task RemoveRoadSegment(ProductContext context, RoadSegmentRemoved roadSegmentRemoved)
     {
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentRemoved.Id);
+        if (roadSegmentRecord == null)
+        {
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentRemoved.Id} is not found");
+        }
 
         context.RoadSegments.Remove(roadSegmentRecord);
     }
